# backend/docs/swagger.yaml
basePath: /
definitions:
  handler.CreateDeviceRequest:
    properties:
      device_code:
        maxLength: 100
        minLength: 1
        type: string
      engine_id:
        maxLength: 100
        type: string
      gsm_number:
        maxLength: 50
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
      serial_number:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - device_code
    - name
    - serial_number
    type: object
...
  handler.HeartbeatRequest:
    properties:
      can_connected:
        type: boolean
      gps_connected:
        type: boolean
      gsm_signal:
        maximum: 0
        minimum: -120
        type: integer
      rs485_connected:
        type: boolean
      sd_card_ok:
        type: boolean
      server_connected:
        type: boolean
      timestamp:
        type: string
    required:
    - gsm_signal
    type: object
...
  handler.UpdateDeviceRequest:
    properties:
      engine_id:
        maxLength: 100
        type: string
      firmware_version:
        maxLength: 50
        type: string
      gsm_number:
        maxLength: 50
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
      status:
        enum:
        - active
        - inactive
        - maintenance
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      errors: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  response.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  service.DeviceCreatedOutput:
    properties:
      device_token:
        type: string
      id:
        type: string
    type: object
  service.DeviceDetail:
    properties:
      created_at:
        type: string
      device_code:
        type: string
      engine_id:
        type: string
      firmware_version:
        type: string
      gsm_number:
        type: string
      id:
        type: string
      metadata: {}
      name:
        type: string
      serial_number:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  service.DeviceListItem:
    properties:
      created_at:
        type: string
      device_code:
        type: string
      engine_id:
        type: string
      firmware_version:
        type: string
      gsm_number:
        type: string
      id:
        type: string
      name:
        type: string
      serial_number:
        type: string
      updated_at:
        type: string
    type: object
  service.DeviceListOutput:
    properties:
      data:
        items:
          $ref: '#/definitions/service.DeviceListItem'
        type: array
      pagination:
        $ref: '#/definitions/service.PaginationMeta'
    type: object
  service.DeviceStatusOutput:
    properties:
      can_connected:
        type: boolean
      device_id:
        type: string
      gps_connected:
        type: boolean
      gsm_signal:
        type: integer
      is_online:
        type: boolean
      last_seen:
        type: string
      rs485_connected:
        type: boolean
      sd_card_ok:
        type: boolean
      server_connected:
        type: boolean
      updated_at:
        type: string
    type: object
...
  service.PaginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@genset-monitoring.io
    name: API Support
  description: Production-ready IoT Genset Monitoring backend API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Genset Monitoring API
  version: "1.0"
paths:
...
  /api/v1/devices:
    get:
      description: Returns a paginated list of devices with optional search, status,
        and online/offline filter
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: Search by device_code or name
        in: query
        name: search
        type: string
      - description: 'Filter by status: active | inactive | maintenance'
        in: query
        name: status
        type: string
      - description: 'Filter by connectivity: true | false (DEPRECATED)'
        in: query
        name: online
        type: string
      - description: 'Sort column: created_at | name | device_code | firmware_version'
        in: query
        name: sort_by
        type: string
      - description: 'Sort direction: asc | desc (default: desc)'
        in: query
        name: sort_dir
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Device list
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.DeviceListOutput'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List Devices
      tags:
      - devices
    post:
      consumes:
      - application/json
      description: Registers a new device; returns its UUID and a one-time device
        token
      parameters:
      - description: Create device payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateDeviceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Device created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.DeviceCreatedOutput'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: device_code or serial_number already in use
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Device
      tags:
      - devices
  /api/v1/devices/{deviceID}:
    delete:
      description: Soft-deletes a device by UUID, preserving the audit trail
      parameters:
      - description: Device UUID
        in: path
        name: deviceID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Device deleted
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Device not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Device
      tags:
      - devices
    get:
      description: Returns full detail of a device by its UUID
      parameters:
      - description: Device UUID
        in: path
        name: deviceID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Device detail
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.DeviceDetail'
              type: object
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Device not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Device Detail
      tags:
      - devices
    patch:
      consumes:
      - application/json
      description: Partially updates device fields (name, firmware_version, status)
      parameters:
      - description: Device UUID
        in: path
        name: deviceID
        required: true
        type: string
      - description: Update device payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Device updated
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid UUID or request body
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Device not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Device
      tags:
      - devices
...
  /api/v1/devices/{deviceID}/heartbeat:
    post:
      consumes:
      - application/json
      description: Receives health check from device and updates last_online_at timestamp
      parameters:
      - description: Device UUID
        in: path
        name: deviceID
        required: true
        type: string
      - description: Heartbeat payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.HeartbeatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Heartbeat recorded
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid payload
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Device not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Device Heartbeat
      tags:
      - status
  /api/v1/devices/{deviceID}/status:
    get:
      description: Returns real-time hardware health and connectivity status
      parameters:
      - description: Device UUID
        in: path
        name: deviceID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Status retrieved
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.DeviceStatusOutput'
              type: object
        "404":
          description: Status not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Device Status
      tags:
      - status
...
securityDefinitions:
  BearerAuth:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer eyJhbGciO..."
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"